+++ ParseCommon/Internal/Modules/IParseModule.cs
@@ -1,0 +1,8 @@
+using System;
+
+namespace Parse.Common.Internal {
+  public interface IParseModule {
+    void OnModuleRegistered();
+    void OnParseInitialized();
+  }
+}
