+++ ParseCommon/Properties/AssemblyInfo.cs
@@ -1,0 +1,20 @@
+// Copyright (c) 2015-present, Parse, LLC.  All rights reserved.  This source code is licensed under the BSD-style license found in the LICENSE file in the root directory of this source tree.  An additional grant of patent rights can be found in the PATENTS file in the same directory.
+
+using System;
+using System.Reflection;
+using System.Runtime.InteropServices;
+using Parse.Common.Internal;
+
+// General Information about an assembly is controlled through the following
+// set of attributes. Change these attribute values to modify the information
+// associated with an assembly.
+[assembly: AssemblyTitle("Parse")]
+[assembly: AssemblyDescription("Makes accessing services from Parse native and straightforward.")]
+[assembly: AssemblyConfiguration("")]
+[assembly: AssemblyCompany("")]
+[assembly: AssemblyProduct("Parse")]
+[assembly: AssemblyCopyright("Copyright © Parse 2012")]
+[assembly: AssemblyTrademark("")]
+[assembly: AssemblyCulture("")]
+
+[assembly: ComVisible(true)]
